✨  errorStaus 페이지 넘버 관련하여 추가

+++ src/main/java/com/ssmoker/smoker/domain/notice/dto/NoticeResponse.java
@@ -19,7 +19,7 @@ public class NoticeResponse {
         List<NoticeViewResponse> notices;
         Integer size;
         Integer totalPage;
-        Integer totalElements;
+        Long totalElements;
         Boolean isFirst;
         Boolean isLast;
     }
